bau: Loosen test that was failing on some machines but not Jenkins
There seems to be some difference in the exception messages thrown by
the certificate timestamp checker depending on the system. It's probably
to do with the version of java.

The test is too strict (do we actually care about the exact exception
message thrown?), so let's just check the reason and not the message.

Authors: @richardtowers