TT-1934: Commonise JSON keys in key parsers
To allow the polymorphic deserialization to work, all subtypes
must use a common key for the value object. This change makes the
keys parsers more idiomatic Jackson and gives them all the
same JSON keys. It also adds many tests for this new functionality

Couple of comments with reference to an outstanding issue with Jackson
SubType parsing that can be avoided by never trying to deserialize a
concrete subtype

Authors: @andy-paine